src/ai/cv/py_requirements.txt
opencv-python==4.5.2.54
numpy==1.21.0
pillow==8.3.0
scipy
scikit-image